+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## Whats New
 
+* 2017-12-18 - PWP - Added wrapper for empty message component
 * 2017-12-15 - CWF - SOHO-6856: Upgraded JQuery types from 2.0.46 to 3.2.16 to match our use of jQuery 3
 * 2017-12-06 - MHH - SOHO-7228: Added support for lazy loading the context menu's popup menu.
 * 2017-12-06 - MAF - SOHO-7216: Added support for SohoAlert.